+++ packages/headless/src/features/attached-results/attached-results-actions-loader.ts
@@ -1,0 +1,44 @@
+import {PayloadAction} from '@reduxjs/toolkit';
+import {InsightEngine} from '../../app/insight-engine/insight-engine';
+import {attachedResults} from '../../app/reducers';
+import {
+  setAttachedResults,
+  SetAttachedResultsActionCreatorPayload,
+} from './attached-results-actions';
+
+export interface AttachedResultsActionCreators {
+  /**
+   * Creates an action that sets the attached results to a record.
+   *
+   * @example
+   *
+   * ```js
+   * const {setAttachedResults} = loadInsightAttachedResultsActions(engine);
+   * ```
+   *
+   * engine.dispatch(setAttachedResults(attachedResults, loading));
+   *
+   * @param attachedResults - The attached results records corresponding to this record.
+   * @param loading - Optional state of loading.
+   * @returns A dispatchable action.
+   */
+  setAttachedResults(
+    payload: SetAttachedResultsActionCreatorPayload
+  ): PayloadAction<SetAttachedResultsActionCreatorPayload>;
+}
+
+/**
+ * Loads the `attachedResults` reducer and returns possible action creators.
+ *
+ * @param engine - The headless engine.
+ * @returns An object holding the action creators.
+ */
+export function loadAttachedResultsActions(
+  engine: InsightEngine
+): AttachedResultsActionCreators {
+  engine.addReducers({attachedResults});
+
+  return {
+    setAttachedResults,
+  };
+}