+++ packages/quantic/force-app/main/default/lwc/quanticResultCopyToClipboard/quanticResultCopyToClipboard.js
@@ -18,6 +18,7 @@ import {LightningElement, api} from 'lwc';
  * <c-quantic-result-copy-to-clipboard engine-id={engineId} result={result} text-template="${title}\n${clickUri}"></c-quantic-result-copy-to-clipboard>
  */
 export default class QuanticResultCopyToClipboard extends LightningElement {
+  static delegatesFocus = true;
   labels = {
     copy,
     copied,
@@ -62,6 +63,8 @@ export default class QuanticResultCopyToClipboard extends LightningElement {
   /** @type {InsightEngine} */
   engine;
 
+  _loading = false;
+
   connectedCallback() {
     this.displayedLabel = this.label;
     registerComponentForInit(this, this.engineId);
@@ -97,20 +100,23 @@ export default class QuanticResultCopyToClipboard extends LightningElement {
    */
   handleCopyToClipBoard = (event) => {
     event.stopPropagation();
-    const {setLoading, result} = event.detail;
-    setLoading(true);
+    const {result} = event.detail;
+    this._loading = true;
     const resultText = buildTemplateTextFromResult(this.textTemplate, result);
 
     copyToClipboard(resultText)
       .then(() => {
-        setLoading(false);
         this.engine.dispatch(this.actions.logCopyToClipboard(this.result));
         this.displayedLabel = this.successLabel;
         this.resetOriginalLabel();
+        // The copy to clipboard fallback method makes the component lose focus, the logic below resets the focus on the button.
+        this.template.host.focus();
       })
       .catch((err) => {
-        setLoading(false);
         console.error('Copy to clipboard action failed.', err);
+      })
+      .finally(() => {
+        this._loading = false;
       });
   };